+++ lr0.c
@@ -9,7 +9,7 @@ extern Yshort *itemset;
 extern Yshort *itemsetend;
 extern unsigned *ruleset;
 
-int nstates;
+int unsigned nstates;
 core *first_state;
 shifts *first_shift;
 reductions *first_reduction;
@@ -36,7 +36,7 @@ static void allocate_itemsets(void)
     register Yshort *itemp;
     register Yshort *item_end;
     register int symbol;
-    register int i;
+    register size_t i;
     register int count;
     register int max;
     register Yshort *symbol_count;
@@ -84,7 +84,7 @@ static void allocate_storage(void)
 
 static core* new_state(int symbol)
 {
-    register int n;
+    register size_t n;
     register core *p;
     register Yshort *isp1;
     register Yshort *isp2;
@@ -127,7 +127,7 @@ static int get_state(int symbol)
     register Yshort *iend;
     register core *sp;
     register int found;
-    register int n;
+    register size_t n;
 
 #ifdef	TRACE
     BtYacc_logf("Entering get_state(%d)\n", symbol);
@@ -183,8 +183,8 @@ static int get_state(int symbol)
 
 static void append_states(void)
 {
-    register int i;
-    register int j;
+    register size_t i;
+    register size_t j;
     register int symbol;
 
 #ifdef	TRACE
@@ -224,7 +224,7 @@ static void free_storage(void)
 
 static void initialize_states(void)
 {
-    register int i;
+    register size_t i;
     register Yshort *start_derives;
     register core *p;
 
@@ -251,7 +251,7 @@ static void initialize_states(void)
 
 static void new_itemsets(void)
 {
-    register int i;
+    register size_t i;
     register int shiftcount;
     register Yshort *isp;
     register Yshort *ksp;
@@ -397,20 +397,20 @@ static void generate_states(void)
 void show_cores()
 {
     core *p;
-    int i, j, k, n;
-    int itemno;
+    int unsigned i, j, k, n;
+    int unsigned itemno;
 
     k = 0;
     for (p = first_state; p; ++k, p = p->next)
     {
 	if (k) printf("\n");
-	printf("state %d, number = %d, accessing symbol = %s\n",
+	printf("state %u, number = %d, accessing symbol = %s\n",
 		k, p->number, symbol_name[p->accessing_symbol]);
 	n = p->nitems;
 	for (i = 0; i < n; ++i)
 	{
 	    itemno = p->items[i];
-	    printf("%4d  ", itemno);
+	    printf("%4u  ", itemno);
 	    j = itemno;
 	    while (ritem[j] >= 0) ++j;
 	    printf("%s :", symbol_name[rlhs[-ritem[j]]]);
@@ -431,20 +431,20 @@ void show_cores()
 
 void show_ritems()
 {
-    int i;
+    int unsigned i;
 
     for (i = 0; i < nitems; ++i)
-	printf("ritem[%d] = %d\n", i, ritem[i]);
+	printf("ritem[%u] = %d\n", i, ritem[i]);
 }
 
 
 /* show_rrhs is used for debugging */
 void show_rrhs()
 {
-    int i;
+    int unsigned i;
 
     for (i = 0; i < nrules; ++i)
-	printf("rrhs[%d] = %d\n", i, rrhs[i]);
+	printf("rrhs[%u] = %d\n", i, rrhs[i]);
 }
 
 
@@ -453,13 +453,13 @@ void show_rrhs()
 void show_shifts()
 {
     shifts *p;
-    int i, j, k;
+    int unsigned i, j, k;
 
     k = 0;
     for (p = first_shift; p; ++k, p = p->next)
     {
 	if (k) printf("\n");
-	printf("shift %d, number = %d, nshifts = %d\n", k, p->number,
+	printf("shift %u, number = %d, nshifts = %d\n", k, p->number,
 		p->nshifts);
 	j = p->nshifts;
 	for (i = 0; i < j; ++i)
@@ -471,7 +471,7 @@ void show_shifts()
 #ifdef	DEBUG
 static void print_derives(void)
 {
-    register int i;
+    register size_t i;
     register Yshort *sp;
 
     printf("\nDERIVES\n\n");
@@ -493,8 +493,8 @@ static void print_derives(void)
 
 static void set_derives(void)
 {
-    register int i, k;
-    register int lhs;
+    register size_t i, k;
+    register size_t lhs;
     register Yshort *rules;
 
     derives = NEW2(nsyms, Yshort *);
@@ -530,7 +530,7 @@ void free_derives()
 
 static void set_nullable(void)
 {
-    register int i, j;
+    register size_t i, j;
     register int empty;
     int done;
