+++ warshall.c
@@ -2,7 +2,7 @@
 
 void transitive_closure(unsigned *R, int n)
 {
-    register int rowsize;
+    register size_t rowsize;
     register unsigned mask;
     register unsigned *rowj;
     register unsigned *rp;
@@ -53,7 +53,7 @@ void transitive_closure(unsigned *R, int n)
 
 void reflexive_transitive_closure(unsigned *R, int n)
 {
-    register int rowsize;
+    register size_t rowsize;
     register unsigned mask;
     register unsigned *rp;
     register unsigned *relend;
